# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Midanic SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout and revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "204":
          description: Logged out successfully

  # ── PRODUCTS ─────────────────────────────────────────────────────────────────
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all published products
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: featured
          schema:
            type: boolean
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{slug}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List published blog posts
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogListResponse"
  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── NEWS ─────────────────────────────────────────────────────────────────────
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List published news articles
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: News articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsListResponse"
  /news/{slug}:
    get:
      operationId: getNewsArticle
      tags: [news]
      summary: Get a news article by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── CONTACT & INQUIRIES ────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RegisterInput:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        companyName:
          type: string
        language:
          type: string
          enum: [en, fr, ar]
          default: en
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    TokenResponse:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string

    # ── USER/PROFILE SCHEMAS ─────────────────────────────────────────────────
    UserProfile:
      type: object
      required: [id, email, firstName, lastName, role, language, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [super_admin, admin, support, billing, customer]
        language:
          type: string
          enum: [en, fr, ar]
        companyName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        twoFactorEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Product:
      type: object
      required: [id, name, slug, description, category, published, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: ["string", "null"]
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        published:
          type: boolean
        trialDays:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    ProductDetail:
      type: object
      required: [id, name, slug, description, category, published, createdAt, versions, downloads]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: ["string", "null"]
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        published:
          type: boolean
        trialDays:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        versions:
          type: array
          items:
            $ref: "#/components/schemas/ProductVersion"
        downloads:
          type: array
          items:
            $ref: "#/components/schemas/DownloadFile"
    ProductVersion:
      type: object
      required: [id, productId, version, releaseNotes, releasedAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        version:
          type: string
        releaseNotes:
          type: ["string", "null"]
        isLatest:
          type: boolean
        releasedAt:
          type: string
          format: date-time
    DownloadFile:
      type: object
      required: [id, productId, fileName, fileSize, platform, downloadUrl, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        versionId:
          type: ["integer", "null"]
        fileName:
          type: string
        fileSize:
          type: integer
        platform:
          type: string
          enum: [windows, mac, linux, android, ios, web]
        version:
          type: ["string", "null"]
        downloadUrl:
          type: string
        downloadCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    # ── BLOG SCHEMAS ─────────────────────────────────────────────────────────
    BlogPost:
      type: object
      required: [id, title, slug, content, published, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        authorName:
          type: ["string", "null"]
        published:
          type: boolean
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    BlogListResponse:
      type: object
      required: [posts, total, page, limit]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── NEWS SCHEMAS ─────────────────────────────────────────────────────────
    NewsArticle:
      type: object
      required: [id, title, slug, content, published, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        published:
          type: boolean
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    NewsListResponse:
      type: object
      required: [articles, total, page, limit]
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/NewsArticle"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── CONTACT SCHEMAS ───────────────────────────────────────────────────────
...